requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning and ML
transformers>=4.20.0
diffusers>=0.20.0
accelerate>=0.20.0
wandb>=0.15.0
tensorboard>=2.10.0

# Robotics and 3D
robotics-toolbox-python>=1.0.0
spatialmath-python>=1.0.0
trimesh>=3.15.0
open3d>=0.16.0

# Data processing
h5py>=3.7.0
zarr>=2.12.0
opencv-python>=4.6.0
Pillow>=9.0.0
tqdm>=4.64.0

# Configuration and utilities
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0
click>=8.0.0
rich>=12.0.0

# Visualization
plotly>=5.10.0
dash>=2.6.0
ipywidgets>=8.0.0
jupyter>=1.0.0

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0

# GFlowNets specific
torch-geometric>=2.1.0
networkx>=2.8.0

# Optional: for real robot integration
# pyrealsense2>=2.50.0
# robotiq-gripper>=0.1.0